services/signal-service/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
pandas
numpy
python-dotenv
pydantic
